build: error out when missing elftools python module
In the case where we use the meson python "find_installation()" function
to get our python binary, we can fail the configure/setup step if the
elftools module is missing. This avoids later errors on build when the
module is missed.

Old output (error logged and config continues):

  Program python3 (elftools) found: NO

New output:
  Program python3 found: YES (/usr/bin/python3)
  Program python3 (elftools) found: NO

  ../buildtools/meson.build:15:31: ERROR: python3 is missing modules: elftools

Signed-off-by: Bruce Richardson <bruce.richardson@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>