net/hns3: VF support parse max TC number
The mailbox message HNS3_MBX_GET_BASIC_INFO can obtain the maximum
number of TCs of the device. The VF does not support multiple TCs,
therefore, this field is not saved.

Now the VF needs to support multiple TCs, therefore, this field needs
to be saved.

This commit also support dump the TC info.

Signed-off-by: Chengwen Feng <fengchengwen@huawei.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>