net/dpaa: add devargs for enabling err packets on main queue
Currently, error queue is mapped to the Rx queue and enabled by default.
This patch adds the devargs to control the err packets on main queue.
Also, in VSP mode the error queue should be disabled because the error
packets from kernel are diverted to the Rx queue/err queue causing crash.

Signed-off-by: Vanshika Shukla <vanshika.shukla@nxp.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>