net/bonding: avoid RSS reta update in flow-isolation mode
In bonding PMD, member_start() function checks whether RSS(mq_mode) is enabled
and then calling the RSS rte_eth_dev_rss_reta_update() API, which is returning
error in-case of device configured in flow isolation-mode. When the device
configured with flow isolation mode RSS reta update is not required and the API
was not supported. For example in-case of mlx5 PMD, RSS reta update API was not
supported, when the device is configured in flow isolatio-mode.

So, added check to verify if the device is configured in flow isolation-mode
then, do not call the rte_eth_dev_rss_reta_update() API.

Signed-off-by: Madhuker Mythri <madhuker.mythri@oracle.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>