bus/auxiliary: fix auxiliary device port destruction
DPDK auxiliary bus first creates a list of known devices in the
`auxiliary_bus.device_list` and then tries to attach a driver to each
device on that list.

If the driver attachment has failed, a device will remain on the list.

The device destruction did not validate driver existence and
crashed.

The patch validates auxiliary device driver before accessing
driver related resources.

Fixes: 1afce3086cf4 ("bus/auxiliary: introduce auxiliary bus")
Cc: stable@dpdk.org

Signed-off-by: Gregory Etelson <getelson@nvidia.com>
Acked-by: Dariusz Sosnowski <dsosnowski@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>