doc: clarify VLAN and QinQ stripping behaviour
The behaviour of VLAN tag stripping Rx offloads is unclear in DPDK, and
not very well documented. Even the documentation that does exist appears
contradictory.

For example, the doxygen docs for the mbuf flag
RTE_MBUF_F_RX_QINQ_STRIPPED says:

  "If RTE_MBUF_F_RX_QINQ_STRIPPED is set and RTE_MBUF_F_RX_VLAN_STRIPPED
  is unset, only the outer VLAN is removed from packet data,..."

but the docs for RTE_MBUF_F_RX_QINQ says:

  "If the flag RTE_MBUF_F_RX_QINQ_STRIPPED is also present, both VLANs
  headers have been stripped from mbuf data, ..."

Without a good definition of what the correct behaviour is, it's not
possible to assess and ensure conformance across drivers. Update the
documentation for NIC features, ethdev and mbuf library to all report
the same information: that VLAN strip feature is stripping one flag, and
QinQ strip feature is removing two.

Summary of VLAN and QinQ stripping behaviour as reported in docs after
this patch:

+-------------------+----------------------+----------------------------+
| Input Traffic     | VLAN-strip on        | QinQ strip on              |
+===================+======================+============================+
| Single VLAN pkts  | Tag in vlan_tci      | Tag in vlan_tci            |
+-------------------+----------------------+----------------------------+
| Double VLAN pkts  | Outer tag in vlan_tci| Outer tag in vlan_tci_outer|
|                   |                      | Inner tag in vlan_tci      |
+-------------------+----------------------+----------------------------+

Signed-off-by: Bruce Richardson <bruce.richardson@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>