eal: define the EAL parameters in argparse format
Create eal_option_list.h, containing all the possible EAL parameters,
and basic info about them, such as type, whether they take a parameter
or not. Each entry is defined using a macro, which will be then
interpreted when the file is included.

First time this header in included in the eal_common_options.c file, the
macros are defined in such a way as to define field elements for an
"eal_init_args" structure, where each value is either a string type, if
it takes a parameter, or boolean type if it doesn't. For those elements
that take multiple values, i.e. are passed multiple times, we put them
in a TAILQ.

The second time of inclusion, the macros are defined so as to define the
arguments in an rte_argparse structure for EAL. For the basic string and
boolean types, we just store the values in the appropriate field in the
previous defined "eal_init_args" structure. For the list elements, we
use the argparse callback to process those elements, adding them to the
TAILQ as they are encountered.

Signed-off-by: Bruce Richardson <bruce.richardson@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>