eal: add warnings about ignored options
The lcoreid-base option only applies when used with the lcores-remapped
option, so print a warning when it's used with an lcores or coremask
flag, or when no core specification is given at all.

To do this, we add some generic infrastructure to allow us to warn about
options which are ignored when either used together or not used
together.

Signed-off-by: Bruce Richardson <bruce.richardson@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>