graph: add callback in dispatch mode
In dispatch mode, the assumption is that the graphs will be
polled continuously: there is no way to know when packets are transferred
from one graph to another.

Introduce a callback to be notified when packets are enqueue in the work
queue: this can be useful if we want to wake up a thread.
This sleep/wake up mechanism is deferred to the application.

Signed-off-by: Christophe Fontaine <cfontain@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>