dts: call hugepages configure at testrun setup
The node class contains a _configure_hugepages method to be called
on the sut node if the user has specified a hugepages config value
in their nodes configuration. However, the commit referenced below
removed the code path to this function, making it unreachable. If
the user has configured hugepages already via kernel parameters
this is harmless, but if they haven't, this bug will cause the DTS
execution to fail. This commit adds a call to configure_hugepages
in the testrun setup, and necessarily makes configure_hugepages a
public method.

Fixes: d77a4a27bd24 ("dts: remove node distinction")

Signed-off-by: Patrick Robb <probb@iol.unh.edu>
Signed-off-by: 0-day Robot <robot@bytheb.org>