dts: enable port binding on the TG
Currently, driver binding for all nodes relies on a devbind_script_path
attribute in the node main sessions. This attribute is being set on the
SUT node, but not the TG node. Consequently, if the TG node ports are
not bound to the correct driver before DTS execution, the TG ports will
fail to bind and DTS will error.

relocate prepare devbind script method to topology class to be used by both
TG and SUT nodes.

Signed-off-by: Andrew Bailey <abailey@iol.unh.edu>

done

Signed-off-by: 0-day Robot <robot@bytheb.org>