bus/vmbus: add device argument to configure if NUMA information on the device should be ignored
In most cases, netvsc is used with a VF device. The application generally
runs with better performance when all the device memory is allocated on
VF's NUMA node, as the VF device carries most of the data packets.

But sometimes netvsc may run on a different NUMA node than that of the VF.
This patch adds a device argument "numa" to allow the application to
configure if this vmbus device (netvsc) should be NUMA aware. The default
behavior is NUMA aware. Setting "numa=0" specifies this vmbus device is
not NUMA aware.

The device argument is parsed at the time of vmbus device probe to ensure
all the following memory allocations are done on the configured NUMA node.

Signed-off-by: Long Li <longli@microsoft.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>