ethdev: hide internal ethtool symbol
No code call this symbol out of ethdev, remove it from exported symbols
and rename it.

Note: there is nothing particularly wrong with exporting this symbol.
It is just unneeded, and this symbol referencing a linux kernel enum
triggers a abidiff warning in the RHEL CI when this enum gets updated.

Signed-off-by: David Marchand <david.marchand@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>