devtools: ignore new warnings on reported-by tag
There were 2 changes introduced in checkpatch.pl script [1][2]
which verify if Reported-by tag is followed by Link or Closes tag.
Example of the warning output:

    WARNING:BAD_REPORTED_BY_LINK:
      Reported-by: should be immediately followed by Closes:
      with a URL to the report

Original purpose is to link the patch to the mailing list discussion
and/or a bug report.

Since in DPDK these tags are not used, these warnings are false
positives and can be ignored in devtools/checkpatches.sh.

[1]: https://github.com/torvalds/linux/commit/d7f1d71e5ef630ab9e15b5821d297a9e1a5fa1da
[2]: https://github.com/torvalds/linux/commit/44c31888098a590b8ec5ba37009e5a983f7c4b46

Signed-off-by: Dariusz Sosnowski <dsosnowski@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>