mbuf: no need to reset all fields on reinitialized mbufs
The 'next' and 'nb_segs' fields are already reset on newly allocated
reinitialized mbufs (a.k.a. raw mbufs), so a simpler reset function for
reinitialized mbufs was added.

The 'ol_flags' field must indicate whether the mbuf uses an external
buffer.
Unlike the normal packet mbuf reset function, which gets this information
from the mbuf itself (by masking the RTE_MBUF_F_EXTERNAL flag), the
simpler reset function gets this information from the mempool, thereby
reducing the number of read-modify-writes on the mbuf from two to one.
The remaining read-modify-write on the mbuf in the simpler reset function
is in rte_pktmbuf_reset_headroom(), where the 'buf_len' field is read
before writing the 'data_off' field.
And I am hoping that it will be possible to eliminate that one too.

Signed-off-by: Morten Brørup <mb@smartsharesystems.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>